Bug Fixes and Improvements
A bug in entry_grid.py caused existing_board_cell to be the updated cell instead of the existing cell. Also, candidates should only be added back if the cell is not empty, this was fixed in cell.py. Improvements were made to board completion, allowing the board to turn green when the user inputs the final digit